feat: SystemAdmins receive a token which is valid on the dps-ingest api (#2835)
